small fixes

+++ controllers/properties.controller.js
@@ -5,7 +5,7 @@ const Review = db.review;
 const Reservation = db.reservation
 const Photo = db.photos_property
 
-const { Op, ValidationError, Sequelize } = require("sequelize");
+const { Op, ValidationError, Sequelize, where } = require("sequelize");
 
 // Obtains general information about all properties. Can be filtered by properties type.
 exports.findAll = async (req, res) => {
@@ -24,7 +24,7 @@ exports.findAll = async (req, res) => {
 
       const types = propertyTypes.map(type => type.type_name.toLowerCase());
 
-      if (type & !types.includes(type.toLowerCase())) {
+      if (type && !types.includes(type.toLowerCase())) {
         return res.status(400).json({
           success: false,
           error: "Invalid type values",
@@ -180,7 +180,7 @@ exports.createProperty = async (req, res) => {
         beds: req.body.beds
       });
 
-      if(req.body.photos && req.body.photos.length > 0){
+      if (req.body.photos && req.body.photos.length > 0) {
         for (let photo of req.body.photos) {
           await Photo.create({
             property_ID: newProperty.ID,
@@ -231,6 +231,7 @@ exports.createProperty = async (req, res) => {
   }
 };
 
+// Allows user to edit their property.
 exports.editProperty = async (req, res) => {
   try {
     let property = await Property.findByPk(req.params.idP);
@@ -241,7 +242,7 @@ exports.editProperty = async (req, res) => {
       });
     }
 
-    if(req.loggedUserId == property.owner_username){
+    if (req.loggedUserId == property.owner_username) {
       if (!req.body || Object.keys(req.body).length == 0) {
         return res.status(400).json({
           success: false,
@@ -310,18 +311,36 @@ exports.editProperty = async (req, res) => {
 exports.deleteProperty = async (req, res) => {
   try {
     let property = await Property.findOne({ where: { ID: req.params.idP } });
+    if (!property) {
+      return res.status(404).json({
+        success: false,
+        error: "Property not found",
+        msg: `The specified property ID does not exist.`,
+      });
+    }
+
+    let dates = 0
+    const currentDate = new Date();
+    let reservationProperty = await Reservation.findAll({ where: { property_ID: req.params.idP } })
+    reservationProperty.forEach((resProp) => {
+      const dateIn = new Date(resProp.dateIn);
+      if (dateIn > currentDate) {
+        dates += 1
+      }
+    });
+
     if (req.loggedUserId == property.owner_username || req.loggedUserRole == 'admin') {
-      let deleteProperty = await Property.destroy({ where: { ID: req.params.idP } });
-      if (deleteProperty == 1) {
+      if (dates == 0) {
+        await Property.destroy({ where: { ID: req.params.idP } });
         return res.json({
           success: true,
           msg: `Property deleted successfully.`,
         });
-      } else {
-        return res.status(404).json({
+      }else{
+        return res.status(403).json({
           success: false,
-          error: "Property not found",
-          msg: `The specified property ID does not exist.`,
+          error: "Forbidden",
+          msg: "You have reservations arranged wait for them to end to delete the building.",
         });
       }
     } else {
@@ -424,9 +443,17 @@ exports.createReview = async (req, res) => {
       });
     }
 
-    await Review.create({ 
-      username: req.loggedUserId, rating: req.body.rating, 
-      comment: req.body.comment, reservation_ID: reservation.ID});
+    if (!req.body.comment || !req.body.rating) {
+      return res.status(400).json({
+        success: false,
+        msg: "Comment and rating are required."
+      });
+    }
+
+    await Review.create({
+      username: req.loggedUserId, rating: req.body.rating,
+      comment: req.body.comment, reservation_ID: reservation.ID
+    });
 
     return res.json({
       success: true,
@@ -459,34 +486,6 @@ exports.createReview = async (req, res) => {
   }
 };
 
-exports.updateReview = async (req, res) => {
-  try {
-    // Logic to update a amenity
-    res
-      .status(200)
-      .json({ success: true, message: "Review updated successfully" });
-  } catch (error) {
-    if (error instanceof Sequelize.ConnectionError) {
-      res.status(503).json({
-        error: "Database Error",
-        msg: "There was an issue connecting to the database. Please try again later"
-      });
-    } else {
-      if (error instanceof Sequelize.ConnectionError) {
-        res.status(503).json({
-          error: "Database Error",
-          msg: "There was an issue connecting to the database. Please try again later"
-        });
-      } else {
-        res.status(500).json({
-          error: "Server Error",
-          msg: "An unexpected error occurred. Please try again later."
-        });
-      }
-    }
-  }
-};
-
 // Handles deletion of a sent review (authentication token must be provided in header).
 exports.deleteReview = async (req, res) => {
   try {
@@ -495,6 +494,16 @@ exports.deleteReview = async (req, res) => {
       where: { property_ID: req.params.idP },
       raw: true
     });
+
+    let property = await Property.findOne({ where: { ID: req.params.idP } })
+
+    if (!property) {
+      return res.status(404).json({
+        success: false,
+        error: "Property not found",
+        msg: "The specified Property ID does not exist."
+      });
+    }
 
     const reservationsFound = reservations.map(reservation => reservation.ID);
